[REFACTOR]: 쿠키 처리 공통화(common-security)
